Hardened Logistics AI Pipeline: Comprehensive Audit and Bug Fixes
- Resolved NameError in AI controller regarding undefined weather state.
- Fixed Pydantic schema type mismatch for travel_time_min (int -> float).
- Unified traffic_level defaults to 0.0 across all prediction paths.
- Synchronized frontend What-If simulator formulas with backend heuristic engine.
- Fixed frontend data parsing for injected tactical scenarios.
- Enriched simulation payload with missing context (origin, destination, news).
- Optimized API Gateway route mounting sequence for cold-start reliability.
- Finalized end-to-end logic validation for production readiness.

+++ backend/ai-service/app/models/schemas.py
@@ -25,7 +25,7 @@ class OptimizationData(BaseModel):
 class ScoredRoute(BaseModel):
     summary: str
     distance_km: float
-    travel_time_min: int
+    travel_time_min: float  # Float: includes float delay prediction addend
     total_cost: float
     total_fuel: float
     risk_level: str
